\name{iowaNW00}
\alias{iowaNW00}
\docType{data}
\title{Iowa NDVI}
\description{
Normalized difference vegetation index (NDVI) values derived from satellite
image data from northwest Iowa in July 2000.  These are 792 values, 
representing NDVI at each pixel on a rectangle with 33 rows and 24 columns.
These data are a small subset of the Global Inventory Modeling and Mapping Studies (GIMMS) datasets available at http://www.glcf.umd.edu/data/gimms/.
}
\usage{data(iowaNW00)}
\format{
  A vector of 792 integer values.  The data are in row-major order. 
}
\references{
Pinzon, J., Brown, M.E. and Tucker, C.J., 2005. Satellite time series correction of orbital drift artifacts using empirical mode decomposition. In: N. Huang (Editor), Hilbert-Huang Transform: Introduction and Applications, pp. 167-186.

Tucker, C.J., J. E. Pinzon, M. E. Brown, D. Slayback, E. W. Pak, R. Mahoney, E. Vermote and N. El Saleous (2005), An Extended AVHRR 8-km NDVI Data Set Compatible with MODIS and SPOT Vegetation NDVI Data. International Journal of Remote Sensing, Vol 26:20, pp 4485-5598.
}


\keyword{datasets}
